Move makeLoggedOutLayout() to client/controller
